==> code/00_initialize.R <==
library(kevostools)
library(modelsummary)
library(fixest)
file.remove(".Renviron")
library(sf)
library(mediation)
library(rgdal)
library(kableExtra)
library(car)
library(rgeos)
library(readxl)
library(classInt)
library(splitstackshape)
library(tigris)
library(cowplot)
library(tidycensus)
library(tidyverse)
library(data.table)
library(ggeffects)

save <- c("cleanup", "theme_bc", "save")
options("modelsummary_format_numeric_latex" = "plain")

cleanup <- function(...){
  save2 <- c(save, ...)
  rm(list=ls(envir = .GlobalEnv)[! ls(envir = .GlobalEnv) %in% save2], envir = .GlobalEnv)
  gc()
}
